Bug Fixes
Fixed "if tz:" in 'set_time_zn()', to "if type(tz) != bool:". This now allows for the timezone to be equal to 0(gmt, utc+0, ect...), as previously, the if statement would detect any "0" as not being a valid timezone.

The second fix was of the "if old_tz:" statement, also within 'set_time_zn()'. It was one indentation too low, and was liked up to the invalid timezone "else" statement. I simply indented it once more.

---END---